dom/imported-tests/webapps/WebStorage/tests/submissions/Infraware/test_storage_local_security.html

<!DOCTYPE HTML>
<html>
 <head>
  <title>Web Storage</title>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
 </head>
 <body>
    <h1>storage_local_security</h1>
	<iframe id="frame" src="iframe/local_security_iframe.html" style="width: 0px; height: 0px;"></iframe>
    <div id="log"></div>
    <script>
		var t1 = async_test('storage local security test');

		iframeWindow = document.getElementById('frame').contentWindow;

		setTimeout(function(){
			try {
				var errFlag =iframeWindow.localStorage.getItem("Security");
				t1.step(function() {
					assert_equals(errFlag, "true", 'SECURITY_ERR error is not raised.')
				});
			} catch (e) {
				t1.step(function() {
					assert_unreached('Error is raised.');
				});
			}
			t1.done();
		}, 500);
    </script>
 </body>
</html>
